+++ Runtime/Purchasables/PurchasableItemSO.cs
@@ -12,7 +12,7 @@ namespace THEBADDEST.VirtualCurrencySystem
 	public class PurchasableItemSO : ScriptableObject, IPurchasableItem
 	{
 		[Header("Purchase Settings")]
-		[SerializeField] private CurrencyType currencyType = CurrencyType.Coin;
+		[SerializeField] private string currencyName = CurrencyType.Coin;
 		[SerializeField] private float price = 0f;
 
 		[Header("Presentation (Optional)")]
@@ -23,9 +23,9 @@ namespace THEBADDEST.VirtualCurrencySystem
 		private Action onSuccess, onFail;
 
 		/// <summary>
-		/// The currency type required for this purchase.
+		/// The currency name required for this purchase.
 		/// </summary>
-		public CurrencyType CurrencyType => currencyType;
+		public string CurrencyName => currencyName;
 
 		/// <summary>
 		/// The price represented as a <see cref="BigNumber"/>.